Rename "iface" package
