+++ src/main/java/appeng/menu/implementations/PatternAccessTermMenu.java
@@ -50,8 +50,8 @@ import appeng.core.AELog;
 import appeng.core.sync.packets.PatternAccessTerminalPacket;
 import appeng.crafting.pattern.EncodedPatternItem;
 import appeng.helpers.InventoryAction;
-import appeng.helpers.iface.PatternProviderLogic;
-import appeng.helpers.iface.PatternProviderLogicHost;
+import appeng.helpers.patternprovider.PatternProviderLogic;
+import appeng.helpers.patternprovider.PatternProviderLogicHost;
 import appeng.menu.AEBaseMenu;
 import appeng.menu.guisync.GuiSync;
 import appeng.parts.crafting.PatternProviderPart;
