+++ src/test/java/appeng/helpers/patternprovider/GenericStackInvTest.java
@@ -1,4 +1,4 @@
-package appeng.helpers.iface;
+package appeng.helpers.patternprovider;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
